added common object id validator
